pyproject.toml
[tool.poetry]
authors = ["None"]
description = ""
name = "noname"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
asyncio = "^3.4.3"
fastapi = "^0.115.4"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}

aiohttp = "^3.11.12"
alembic = "^1.15.1"
asyncpg = "^0.30.0"
gunicorn = "^23.0.0"
httpx = "^0.28.1"
loguru = "^0.7.2"
passlib = "^1.7.4"
prometheus-fastapi-instrumentator = "^7.0.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python = "^3.12"
python-jose = "^3.4.0"
redis = "^4.2.0rc1"
sqladmin = "^0.20.1"
sqlalchemy = "^2.0.39"
types-passlib = "^1.7.7.20250322"
types-python-jose = "^3.4.0.20250224"
uvicorn = "^0.32.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
coverage = "^7.6.2"
flake8-pyproject = "^1.2.3"
isort = "^5.13.2"
line-profiler = "^4.1.3"
locust = "^2.32.2"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-benchmark = "^5.1.0"
pytest-mock = "^3.14.0"
pytest-profiling = "^1.7.0"
ruff = "^0.9.10"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
log_cli = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
exclude = [
  "./src/database/migrations/*",
]
ignore = [
  "C901", # too complex
  "W191", # indentation contains tabs
  "B904", # raise from None
]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "N", # PEP8 naming convetions
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.black]
# https://github.com/psf/black
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''
line-length = 88
target_version = ['py312']

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
force_grid_wrap = 0
include_trailing_comma = true
indent = '    '
line_length = 88
multi_line_output = 3

[tool.coverage]
# https://github.com/nedbat/coveragepy
[tool.coverage.run]

branch = true
omit = ['']
source = ["src"]

[tool.coverage.report]
exclude_lines = ["# noqa", "raise NotImplementedError"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
follow_imports = "silent"
mypy_path = "./src"
no_implicit_reexport = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = false

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "nebula3.*"

[tool.flake8]
count = true
exclude = [
  '.git',
  '__pycache__',
  '__init__.py',
  './src/database/migrations/',
]
ignore = ['E231', 'E241', 'W503']
max-line-length = 88
per-file-ignores = [
  '__init__.py:F401',
  '__init__.py:F403',
  './src/endpoints/schemas.py:E501',
]
